requirements.txt
flask==3.0.3
pillow==10.4.0
numpy==1.24.4
tflite-runtime==2.14.0
gunicorn==21.2.0
